pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "industrial-data-logger"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "paho-mqtt==1.6.1",
    "asyncua==1.1.5",
    "frappeclient==0.0.7",
    "rich>=13.7",
    "python-dotenv>=1.0",
]
